backend/requirements.txt
CurrencyConverter==0.17.14
Django
PyJWT==2.8.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
pip==23.3.2
psycopg2-binary
python-dotenv==1.0.0
stripe==7.10.0
